// src/app/enums/funcionalidade.ts
export enum FuncionalidadeEnum {
    USUARIO = 1,
    PERFIL_USUARIO = 2,
    PESSOA = 3,
    EVENTOS = 4,
    VISITAS = 5,
    LEVANTAMENTOS = 6,
    EMPRESTIMOS = 7,
    EMPRESA_ORGAO = 8,
    ENDERECO = 9,
    TELEFONE = 10,
    PARAMETRO = 11,
    PARAMETRO_BOLETO = 12,
    ARQUIVO_REMESSA = 13,
    LINHA_CREDITO = 14,
    
    EMPRESA_CONSIGNATARIA = 95,
    EMPREGADO_CONSIGNADO = 96,
    CONSULTAR_MARGEM = 97,
    AUTORIZAR_CONSULTA_MARGEM = 98,
    CONSIGNACAO = 911,
    CONVENIO = 912,
    TAXAS = 913,
    SIMULADOR = 914,
    CONSIGNACAO_EXTRATO = 915,
}